requirements.txt
kafka-python>=2.0.2
PyYAML>=5.4.1
python-dateutil>=2.8.2
ipaddress>=1.0.23
tenacity>=8.0.1  # for retry logic